+++ packages/core/android/src/newarch/java/io/sentry/react/RNSentryModule.java
@@ -129,6 +129,11 @@ public class RNSentryModule extends NativeRNSentrySpec {
   }
 
   @Override
+  public void removeAttribute(String key) {
+    this.impl.removeAttribute(key);
+  }
+
+  @Override
   public void closeNativeSdk(Promise promise) {
     this.impl.closeNativeSdk(promise);
   }